% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{mecab_token_df}
\alias{mecab_token_df}
\title{Create token via data.frame}
\usage{
mecab_token_df(data, sentence, sys_dic = "", user_dic = "")
}
\arguments{
\item{data}{data.frame}

\item{sentence}{A character vector of any length. For analyzing multiple sentences, put them in one character vector.}

\item{sys_dic}{A location of system MeCab dictionary. The default value is "".}

\item{user_dic}{A location of user-specific MeCab dictionary. The default value is "".}
}
\description{
Create token via data.frame
}
