% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{ngram_df}
\alias{ngram_df}
\title{N-gram data.frame}
\usage{
ngram_df(data, sentence, ngram = 2L, collapse = " ", ...)
}
\arguments{
\item{data}{data.frame}

\item{sentence}{A character vector of any length. For analyzing multiple sentences, put them in one character vector.}

\item{ngram}{N-gram count. Default is set 2 (bi-gram).}

\item{collapse}{an optional character string to separate the tokens.}

\item{...}{Additional arguments passed to \link[RcppMeCab:pos]{RcppMeCab::pos}}
}
\description{
N-gram data.frame
}
\examples{
\dontrun{
d <- data.frame(x = intToUtf8(c(12377L, 12418L, 12418L, 12418L, 12418L,
                                12418L, 12418L, 12418L, 12418L, 12398L,
                                12358L, 12385L)))
d \%>\%
  ngram_df(x)
# Tri-gram
d \%>\%
  ngram_df(x, n = 3)
# Uni-gram
d \%>\%
  ngram_df(x, n = 1)
}
}
