% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{token_tfidf}
\alias{token_tfidf}
\title{Create tf-idf data.frame}
\usage{
token_tfidf(data, id, token, sort = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{id}{variable name as identifier}

\item{token}{token variable name}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}
}
\value{
\link[tibble:tibble]{tibble}
}
\description{
Create tf-idf data.frame
}
\examples{
\dontrun{
data.frame(x = intToUtf8(c(12377L, 12418L, 12418L, 12418L, 12418L,
12418L, 12418L, 12418L, 12418L, 12398L, 12358L, 12385L))) \%>\%
mecab_token_df(x) \%>\%
dplyr::filter(pos == intToUtf8(c(21517L, 35422L))) \%>\%
token_tfidf(doc_id, token)
}
}
