% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{mecab_token_df}
\alias{mecab_token_df}
\title{Create token via data.frame}
\usage{
mecab_token_df(data, sentence, ...)
}
\arguments{
\item{data}{data.frame}

\item{sentence}{A character vector of any length. For analyzing multiple sentences, put them in one character vector.}

\item{...}{Additional arguments passed to \link[RcppMeCab:pos]{RcppMeCab::pos}}
}
\value{
\link[tibble:tibble]{tibble}
}
\description{
Create token via data.frame
}
\examples{
\dontrun{
data.frame(x = intToUtf8(c(12377L, 12418L, 12418L, 12418L, 12418L,
12418L, 12418L, 12418L, 12418L, 12398L, 12358L, 12385L))) \%>\%
  mecab_token_df(x)
}
}
