% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{token_tfidf}
\alias{token_tfidf}
\title{Create tf-idf data.frame}
\usage{
token_tfidf(data, id, token, sort = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{id}{variable name as identifier}

\item{token}{token variable name}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}
}
\description{
Create tf-idf data.frame
}
